+++ scripts/cp4ba-install-db.sh
@@ -38,7 +38,9 @@ resourceExist () {
   return 1
 }
 
-
+#--------------------------------
+# Cluster EDB (deprecated)
+#--------------------------------
 _deployDBClusterEDB () {
 
   if [[ ! -z "$1" ]] && [[ ! -z "$2" ]]; then
@@ -172,7 +174,14 @@ EOF
   fi
 }
 
-_deployDBClusterOSS () {
+#--------------------------------
+# Postgresql (StatefulSet)
+#--------------------------------
+
+#--------------------------------
+# Postgres instance for all databases but ZEN,BTS,IM
+
+_deployPostgresNoSSL () {
 
   if [[ ! -z "$1" ]] && [[ ! -z "$2" ]]; then
 
@@ -187,6 +196,23 @@ _deployDBClusterOSS () {
     echo "  CP4BA '${CP4BA_INST_APPVER}' use image name: "${_imageName}
 
     _PG_SS_CR_TMP="/tmp/cp4ba-pg-statefulset-$USER-$RANDOM"
+
+    _PG_CONFIG_CM=my-postgresql-config    
+    _PG_CONF_FOLDER=/tmp/cp4ba-pg-conf-folder-$USER-$RANDOM
+    mkdir -p ${_PG_CONF_FOLDER} 2>/dev/null 1>/dev/null
+
+    # copiare template .conf
+    if [[ -f "${CP4BA_INST_DB_POSTGRES_CONF_TEMPLATE}"  ]]; then
+      cp ${CP4BA_INST_DB_POSTGRES_CONF_TEMPLATE} ${_PG_CONF_FOLDER}/
+    else
+      echo -e "${_CLR_RED}[✗] ERROR: _deployPostgresNoSSL template not found: ${CP4BA_INST_DB_POSTGRES_CONF_TEMPLATE}${_CLR_NC}"
+      exit 1
+    fi
+
+    # create CM for PG configuration files
+    oc delete configmap -n ${CP4BA_INST_NAMESPACE} ${_PG_CONFIG_CM} 2>/dev/null 1>/dev/null
+    oc create configmap -n ${CP4BA_INST_NAMESPACE} ${_PG_CONFIG_CM} --from-file=${_PG_CONF_FOLDER}/ 2>/dev/null 1>/dev/null
+
 
 cat <<EOF > ${_PG_SS_CR_TMP}
 kind: StatefulSet
@@ -224,6 +250,15 @@ spec:
       labels:
         app: $1
     spec:
+      # 20260223
+      volumes:
+        - name: config-vol
+          configMap:
+            name: ${_PG_CONFIG_CM}
+            defaultMode: 0640
+        - name: data-volume
+          emptyDir: {}
+
       containers:
         - name: postgres
           image: '${_imageName}'
@@ -244,12 +279,27 @@ spec:
           terminationMessagePath: /dev/termination-log
           terminationMessagePolicy: File
           imagePullPolicy: IfNotPresent
+
+          #20260223
+          volumeMounts:
+            - name: config-vol
+              mountPath: /etc/postgresql/
+            - name: data-volume
+              mountPath: /etc/postgresql-data
+                  
+          args:
+            - -c
+            - config_file=/etc/postgresql/postgresql_nossl.conf
+
       restartPolicy: Always
       terminationGracePeriodSeconds: 10
       dnsPolicy: ClusterFirst
-      serviceAccountName: postgres-anyuid
-      serviceAccount: postgres-anyuid
-      securityContext: {}
+      serviceAccountName: ibm-cp4ba-anyuid
+      serviceAccount: ibm-cp4ba-anyuid
+      securityContext:
+        runAsUser: 999
+        runAsGroup: 999   
+        fsGroup: 999
       schedulerName: default-scheduler
   podManagementPolicy: OrderedReady
   updateStrategy:
@@ -307,11 +357,6 @@ spec:
     - IPv4
   internalTrafficPolicy: Cluster
   type: ClusterIP
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: postgres-anyuid
 EOF
 
 
@@ -330,7 +375,7 @@ EOF
         
         # -- if OK end loop
         if [ $? -eq 1 ]; then
-          oc adm policy add-scc-to-user anyuid -z postgres-anyuid -n $2 2>/dev/null 1>/dev/null
+          # oc adm policy add-scc-to-user anyuid -z postgres-anyuid -n $2 2>/dev/null 1>/dev/null
 
           echo "Deployed statefulsets.apps name '$1'"
           break
@@ -343,7 +388,385 @@ EOF
     rm ${_PG_SS_CR_TMP} 2>/dev/null 1>/dev/null
 
   else
-    echo -e "${_CLR_RED}[✗] ERROR: _deployDBClusterOSS name or namespace empty${_CLR_NC}"
+    echo -e "${_CLR_RED}[✗] ERROR: _deployPostgresNoSSL name or namespace empty${_CLR_NC}"
+    exit 1
+  fi
+}
+
+#--------------------------------
+# Postgresql (StatefulSet)
+#--------------------------------
+
+#--------------------------------
+# Postgres instance with SSL enabled (ZEN,BTS,IM external databases)
+
+_createDBCertificates () {
+# $1 = certificate folder
+# $2 = server name
+_CERT_FOLDER=$1
+_CERT_SERVER_NAME=$2
+
+_CERT_PASS=marco
+_CERT_VERIFY="false"
+_CERT_CA_PREFIX_NAME="my-ca"
+_CERT_SERVER_PREFIX_NAME="my-server"
+_CERT_CLIENT_PREFIX_NAME="my-client"
+
+_CERT_ISSUER_SUBJ="/CN=my-postgres"
+_CERT_SUBJ_SERVER="/CN=${_CERT_SERVER_NAME}"
+_CERT_SUBJ_CLIENT="/CN=client.${_CERT_SERVER_NAME}"
+
+# CA conf
+mkdir -p ${_CERT_FOLDER}/ca.db.certs   # Signed certificates storage
+touch ${_CERT_FOLDER}/ca.db.index      # Index of signed certificates
+echo 01 > ${_CERT_FOLDER}/ca.db.serial # Next (sequential) serial number
+
+# Configuration cert server
+cat <<EOF > ${_CERT_FOLDER}/req.conf
+[ req ]
+distinguished_name = req_distinguished_name
+x509_extensions = v3_ca
+prompt = no
+[ req_distinguished_name ]
+C = CN
+ST = MA
+O = CP4BA
+CN = root
+[ v3_ca ]
+basicConstraints = critical,CA:TRUE
+subjectKeyIdentifier = hash
+authorityKeyIdentifier = keyid:always,issuer:always
+[ v3_req ]
+keyUsage = keyEncipherment, dataEncipherment, digitalSignature
+extendedKeyUsage = clientAuth, serverAuth
+subjectAltName = @alt_names
+[ alt_names ]
+IP.1 = 127.0.0.1
+DNS.1 = ${CP4BA_INST_DB_1_CR_NAME_SSL}
+DNS.2 = ${CP4BA_INST_DB_1_SERVICE_SSL}
+DNS.3 = ${CP4BA_INST_DB_1_SERVICE_SSL_R}
+DNS.4 = ${CP4BA_INST_DB_1_SERVER_NAME_SSL}
+DNS.5 = ${CP4BA_INST_DB_1_SERVER_NAME_SSL_R}
+DNS.6 = localhost
+
+EOF
+
+# Configuration cert client
+cat <<EOF > ${_CERT_FOLDER}/req-client.conf
+[ req ]
+distinguished_name = req_distinguished_name
+x509_extensions = v3_ca
+prompt = no
+[ req_distinguished_name ]
+C = CN
+ST = MA
+O = CP4BAClient
+CN = root
+[ v3_ca ]
+basicConstraints = critical,CA:TRUE
+subjectKeyIdentifier = hash
+authorityKeyIdentifier = keyid:always,issuer:always
+[ v3_req ]
+keyUsage = keyEncipherment, dataEncipherment, digitalSignature
+extendedKeyUsage = clientAuth
+
+EOF
+
+# CA
+openssl genrsa -out ${_CERT_FOLDER}/ca.key 2048 2>/dev/null 1>/dev/null
+openssl req -x509 -new -key ${_CERT_FOLDER}/ca.key -sha256 -days 36500 -out ${_CERT_FOLDER}/ca.cert -extensions 'v3_ca' -config ${_CERT_FOLDER}/req.conf 2>/dev/null 1>/dev/null
+
+# SERVER
+openssl genrsa -out ${_CERT_FOLDER}/server.key 2048 2>/dev/null 1>/dev/null
+openssl req -new -sha256 -key ${_CERT_FOLDER}/server.key -out ${_CERT_FOLDER}/server-req.pem -subj "/CN=${CP4BA_INST_DB_1_CR_NAME_SSL}" -config ${_CERT_FOLDER}/req.conf 2>/dev/null 1>/dev/null
+openssl x509 -req -days 36500 -sha256 -extensions v3_req -CA ${_CERT_FOLDER}/ca.cert -CAkey ${_CERT_FOLDER}/ca.key -CAcreateserial -in ${_CERT_FOLDER}/server-req.pem -out ${_CERT_FOLDER}/server.cert -extfile ${_CERT_FOLDER}/req.conf 2>/dev/null 1>/dev/null
+
+# CLIENT
+openssl genrsa -out ${_CERT_FOLDER}/client.key 2048 2>/dev/null 1>/dev/null
+openssl req -new -sha256 -key ${_CERT_FOLDER}/client.key -out ${_CERT_FOLDER}/client-req.pem -subj "/CN=postgres-client" -config ${_CERT_FOLDER}/req-client.conf 2>/dev/null 1>/dev/null
+openssl x509 -req -days 36500 -sha256 -extensions v3_req -CA ${_CERT_FOLDER}/ca.cert -CAkey ${_CERT_FOLDER}/ca.key -CAcreateserial -in ${_CERT_FOLDER}/client-req.pem -out ${_CERT_FOLDER}/client.cert -extfile ${_CERT_FOLDER}/req-client.conf 2>/dev/null 1>/dev/null
+
+}
+
+_deployPostgresSSL () {
+
+  if [[ ! -z "$1" ]] && [[ ! -z "$2" ]]; then
+
+    resourceExist $2 statefulsets.apps $1
+    if [ $? -eq 1 ]; then
+      oc delete statefulsets.apps -n $2 $1 2>/dev/null 1>/dev/null
+    fi
+
+    _imageName="${CP4BA_INST_DB_OSS_IMAGE}"
+
+    echo "Deploying SSL statefulset name '$1'"
+    echo "  CP4BA '${CP4BA_INST_APPVER}' use image name: "${_imageName}
+
+    _PG_SS_CR_TMP="/tmp/cp4ba-pg-statefulset-$USER-$RANDOM"
+
+
+    _PG_CONFIG_CM=my-postgresql-config-ssl
+    _PG_SECRETS=my-postgresql-secrets
+    
+    _PG_CONF_FOLDER=/tmp/cp4ba-pg-conf-folder-$USER-$RANDOM
+    _PG_SECRETS_FOLDER=/tmp/cp4ba-pg-secrets-folder-$USER-$RANDOM
+
+    mkdir -p ${_PG_CONF_FOLDER} 2>/dev/null 1>/dev/null
+    mkdir -p ${_PG_SECRETS_FOLDER} 2>/dev/null 1>/dev/null
+
+    _PG_CONF_FILE=${_PG_CONF_FOLDER}/postgresql_nossl.conf
+    _PG_CONF_FILE_SSL=${_PG_CONF_FOLDER}/postgresql.conf
+    _PG_HBA_CONF_FILE_TMP=${_PG_CONF_FOLDER}/pg_hba.conf.tmp
+    _PG_HBA_CONF_FILE=${_PG_CONF_FOLDER}/pg_hba.conf
+
+    # copiare template .conf
+    if [[ -f "${CP4BA_INST_DB_POSTGRES_CONF_SSL_TEMPLATE}"  ]]; then
+      cp ${CP4BA_INST_DB_POSTGRES_CONF_SSL_TEMPLATE} ${_PG_CONF_FOLDER}/ 2>/dev/null 1>/dev/null
+    else
+      echo -e "${_CLR_RED}[✗] ERROR: _deployPostgresSSL template not found: ${CP4BA_INST_DB_POSTGRES_CONF_SSL_TEMPLATE}${_CLR_NC}"
+      exit 1
+    fi
+    if [[ -f "${CP4BA_INST_DB_POSTGRES_CONF_PGHBA_TEMPLATE}"  ]]; then
+      cp ${CP4BA_INST_DB_POSTGRES_CONF_PGHBA_TEMPLATE} ${_PG_HBA_CONF_FILE_TMP} 2>/dev/null 1>/dev/null
+    else
+      echo -e "${_CLR_RED}[✗] ERROR: _deployPostgresSSL template not found: ${CP4BA_INST_DB_POSTGRES_CONF_PGHBA_TEMPLATE}${_CLR_NC}"
+      exit 1
+    fi
+
+    cat ${_PG_HBA_CONF_FILE_TMP} | sed "s/host all all all scram-sha-256/host all all all trust/g" > ${_PG_HBA_CONF_FILE}
+
+    echo "hostssl all all 127.0.0.1/32 password clientcert=verify-ca" >> ${_PG_HBA_CONF_FILE}
+    echo "hostssl all all 127.0.0.1/32 cert clientcert=verify-full" >> ${_PG_HBA_CONF_FILE}
+
+    # create CM for PG configuration files
+    oc delete configmap -n ${CP4BA_INST_NAMESPACE} ${_PG_CONFIG_CM} 2>/dev/null 1>/dev/null
+    oc create configmap -n ${CP4BA_INST_NAMESPACE} ${_PG_CONFIG_CM} --from-file=${_PG_CONF_FOLDER}/ 2>/dev/null 1>/dev/null
+
+    _createDBCertificates ${_PG_SECRETS_FOLDER} ${CP4BA_INST_DB_1_SERVICE_SSL}
+
+    oc delete secret -n ${CP4BA_INST_NAMESPACE} ${_PG_SECRETS} 2>/dev/null 1>/dev/null
+    oc create secret generic -n ${CP4BA_INST_NAMESPACE} ${_PG_SECRETS} --from-file=${_PG_SECRETS_FOLDER}/ 2>/dev/null 1>/dev/null
+
+cat <<EOF > ${_PG_SS_CR_TMP}
+kind: StatefulSet
+apiVersion: apps/v1
+metadata:
+  name: $1
+  namespace: $2
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: $1
+  serviceName: $1
+  persistentVolumeClaimRetentionPolicy:
+    whenDeleted: Retain
+    whenScaled: Retain
+  volumeClaimTemplates:
+    - kind: PersistentVolumeClaim
+      apiVersion: v1
+      metadata:
+        name: $1
+      spec:
+        accessModes:
+          - ReadWriteOnce
+        resources:
+          requests:
+            storage: ${CP4BA_INST_DB_STORAGE_SIZE}
+        storageClassName: ${CP4BA_SC_NAME}
+        volumeMode: Filesystem
+      status:
+        phase: Pending
+  template:
+    metadata:
+      creationTimestamp: null
+      labels:
+        app: $1
+    spec:
+      volumes:
+        - name: config-vol
+          configMap:
+            name: ${_PG_CONFIG_CM}
+            defaultMode: 0640
+        - name: secret-vol
+          secret:
+            secretName: ${_PG_SECRETS}
+            defaultMode: 0640
+        - name: data-volume
+          emptyDir: {}
+
+      containers:
+        - name: postgres
+
+          image: '${_imageName}'
+          ports:
+            - name: postgres
+              containerPort: 5432
+              protocol: TCP
+          env:
+            - name: POSTGRES_PASSWORD
+              value: "${CP4BA_INST_DB_OSS_ADMIN_PASSWORD}"
+          resources:
+            limits:
+              cpu: "${CP4BA_INST_DB_LIMITS_CPU}"
+              memory: "${CP4BA_INST_DB_LIMITS_MEMORY}"
+            requests:
+              cpu: "${CP4BA_INST_DB_REQS_CPU}"
+              memory: "${CP4BA_INST_DB_REQS_MEMORY}"
+          terminationMessagePath: /dev/termination-log
+          terminationMessagePolicy: File
+          imagePullPolicy: IfNotPresent
+
+          volumeMounts:
+            - name: config-vol
+              mountPath: /etc/postgresql/
+            - name: secret-vol
+              mountPath: /etc/postgresql-secrets
+            - name: data-volume
+              mountPath: /etc/postgresql-data
+                  
+          args:
+            - -c
+            - hba_file=/etc/postgresql/pg_hba.conf
+            - -c
+            - config_file=/etc/postgresql/postgresql_ssl.conf
+
+      restartPolicy: Always
+      terminationGracePeriodSeconds: 10
+      dnsPolicy: ClusterFirst
+      serviceAccountName: ibm-cp4ba-anyuid
+      serviceAccount: ibm-cp4ba-anyuid
+      securityContext:
+        runAsUser: 999
+        runAsGroup: 999   
+        fsGroup: 999
+      schedulerName: default-scheduler
+  podManagementPolicy: OrderedReady
+  updateStrategy:
+    type: RollingUpdate
+    rollingUpdate:
+      partition: 0
+---
+kind: Service
+apiVersion: v1
+metadata:
+  name: $1-rw
+  namespace: $2
+spec:
+  selector:
+    app: $1
+  ports:
+    - protocol: TCP
+      port: 5432
+      targetPort: 5432
+  ipFamilies:
+    - IPv4
+  internalTrafficPolicy: Cluster
+  type: ClusterIP
+---
+kind: Service
+apiVersion: v1
+metadata:
+  name: $1-ro
+  namespace: $2
+spec:
+  selector:
+    app: $1
+  ports:
+    - protocol: TCP
+      port: 5432
+      targetPort: 5432
+  ipFamilies:
+    - IPv4
+  internalTrafficPolicy: Cluster
+  type: ClusterIP
+---
+kind: Service
+apiVersion: v1
+metadata:
+  name: $1-r
+  namespace: $2
+spec:
+  selector:
+    app: $1
+  ports:
+    - protocol: TCP
+      port: 5432
+      targetPort: 5432
+  ipFamilies:
+    - IPv4
+  internalTrafficPolicy: Cluster
+  type: ClusterIP
+EOF
+
+
+    # - loop 
+    while [ true ]
+    do
+      # apply CR
+      oc apply -n $2 -f ${_PG_SS_CR_TMP} 2>/dev/null 1>/dev/null
+      if [ $? -gt 0 ]; then
+        sleep 10
+      else
+        # -- verify CR existence
+        resourceExist $2 "statefulsets.apps" $1
+        
+        # -- if OK end loop
+        if [ $? -eq 1 ]; then
+          echo "Deployed statefulsets.apps name '$1'"
+          break
+        else
+          sleep 1
+        fi
+      fi
+    done
+
+    # https://www.ibm.com/docs/en/cloud-paks/foundational-services/4.x_cd?topic=management-configuring-external-postgresql-database-im
+
+    _SEC_NAME="im-datastore-edb-secret"
+    oc delete secret ${_SEC_NAME} -n "$2" >/dev/null 2>&1
+    oc create secret generic ${_SEC_NAME} \
+      --from-file=ca.crt="${_PG_SECRETS_FOLDER}/ca.cert" \
+      --from-file=tls.crt="${_PG_SECRETS_FOLDER}/client.cert" \
+      --from-file=tls.key="${_PG_SECRETS_FOLDER}/client.key" \
+      --type=kubernetes.io/tls -n "$2" 2>/dev/null 1>/dev/null
+    oc label secret ${_SEC_NAME} cp4ba.ibm.com/backup-type=mandatory -n "$2" 2>/dev/null 1>/dev/null
+
+    _SEC_NAME="ibm-zen-metastore-edb-secret"
+    oc delete secret ${_SEC_NAME} -n "$2" 2>/dev/null 1>/dev/null
+    oc create secret generic -n "$2" ${_SEC_NAME} \
+      --from-file=ca.crt="${_PG_SECRETS_FOLDER}/ca.cert"  \
+      --from-file=tls.crt="${_PG_SECRETS_FOLDER}/client.cert"  \
+      --from-file=tls.key="${_PG_SECRETS_FOLDER}/client.key"  \
+      --type=kubernetes.io/tls 2>/dev/null 1>/dev/null
+    oc label secret -n "$2" ${_SEC_NAME} cp4ba.ibm.com/backup-type=mandatory 2>/dev/null 1>/dev/null
+
+    _SEC_NAME="bts-datastore-edb-secret"
+    openssl pkcs8 -topk8 -inform PEM -outform DER -nocrypt \
+      -in ${_PG_SECRETS_FOLDER}/client.key \
+      -out ${_PG_SECRETS_FOLDER}/tls_key.pk8 2>/dev/null 1>/dev/null
+
+    oc delete secret -n "$2" ${_SEC_NAME} 2>/dev/null 1>/dev/null
+    oc create secret generic -n "$2" ${_SEC_NAME}  \
+      --from-file=ca.crt="${_PG_SECRETS_FOLDER}/ca.cert"  \
+      --from-file=tls.crt="${_PG_SECRETS_FOLDER}/client.cert"  \
+      --from-file=tls.key="${_PG_SECRETS_FOLDER}/tls_key.pk8" 2>/dev/null 1>/dev/null
+    oc label secret -n "$2" ${_SEC_NAME} cp4ba.ibm.com/backup-type=mandatory 2>/dev/null 1>/dev/null
+
+
+    rm ${_PG_SS_CR_TMP} 2>/dev/null 1>/dev/null
+
+    if [[ ! -z "${_PG_CONF_FOLDER}" ]]; then
+      #echo "folder not removed: ${_PG_CONF_FOLDER}"
+      rm -fr ${_PG_CONF_FOLDER} 2>/dev/null 1>/dev/null
+    fi
+    if [[ ! -z "${_PG_SECRETS_FOLDER}" ]]; then
+      #echo "folder not removed: ${_PG_SECRETS_FOLDER}"
+      rm -fr ${_PG_SECRETS_FOLDER} 2>/dev/null 1>/dev/null
+    fi
+
+  else
+    echo -e "${_CLR_RED}[✗] ERROR: _deployPostgresSSL name or namespace empty${_CLR_NC}"
     exit 1
   fi
 }
@@ -352,7 +775,9 @@ _deployDBCluster () {
   if [[ -z "${CP4BA_INST_DB_USE_EDB}" ]] || [[ "${CP4BA_INST_DB_USE_EDB}" = "true" ]]; then
     _deployDBClusterEDB "$1" "$2"
   else
-    _deployDBClusterOSS "$1" "$2"
+    _deployPostgresNoSSL "$1" "$2"
+    sleep 5
+    _deployPostgresSSL "$3" "$2"
   fi
 
 }
@@ -363,7 +788,7 @@ deployDBCluster() {
 # $3: namespace
   if [[ "$1" = "true" ]]; then
     echo -e "${_CLR_GREEN}Deploying DB Cluster '${_CLR_YELLOW}$2${_CLR_GREEN}' in namespace '${_CLR_YELLOW}$3${_CLR_GREEN}'${_CLR_NC}"
-    _deployDBCluster "$2" "$3"
+    _deployDBCluster "$2" "$3" "$4"
   else
     echo -e "${_CLR_YELLOW}Skipping deployment of DB Cluster '${_CLR_GREEN}$1${_CLR_YELLOW}'${_CLR_NC}"
   fi
@@ -448,8 +873,9 @@ deployDBClusters() {
   while [[ $i -le $_IDX_END ]]
   do
     _INST_DB_CR_NAME="CP4BA_INST_DB_"$i"_CR_NAME"
+    _INST_DB_CR_NAME_SSL="CP4BA_INST_DB_"$i"_CR_NAME_SSL"
     if [[ ! -z "${!_INST_DB_CR_NAME}" ]]; then
-      deployDBCluster ${CP4BA_INST_DB} ${!_INST_DB_CR_NAME} $1
+      deployDBCluster ${CP4BA_INST_DB} ${!_INST_DB_CR_NAME} $1 ${!_INST_DB_CR_NAME_SSL}
     else
       echo -e "${_CLR_RED}ERROR, env var '${_CLR_GREEN}${_INST_DB_CR_NAME}${_CLR_RED}' not defined, verify CP4BA_INST_DB_INSTANCES and CP4BA_INST_DB_* values.${_CLR_NC}"
     fi